requirements.txt
arcade==2.5.6
attrs==20.3.0
cffi==1.14.5
numpy==1.20.2
Pillow==8.2.0
pycparser==2.20
pyglet==1.5.15
pymunk==6.0.0
pytiled-parser==0.9.4a3
PyYAML==5.4.1
Shapely==1.7.1
typing-extensions==3.7.4.3
fsmpy == 2.0.0

